Bruh

+++ megumin/modulos/afk.py
@@ -18,13 +18,16 @@ async def rem_afk(c: megux, m: Message):
 
     user_afk = await AFK_STATUS.find_one({"_afk": "on"})
 
+    if m.text == "/afk" or "brb":
+        return 
+    else: 
 
-    if not user_afk:
-        return
-    else:
+        if not user_afk:
+            return
+        else:
 
-        await AFK_STATUS.drop()
-        await m.reply_text(f"{m.from_user.first_name} não está mais AFK!")
+            await AFK_STATUS.drop()
+            await m.reply_text(f"{m.from_user.first_name} não está mais AFK!")
 
 
 @megux.on_message(filters.command("afk"))
